+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changelog
 These are the changes brought in each versions:
 
-v1.0.5 (TBD)
+v1.0.5 (2020-09-16)
 ===================
 Non-breaking changes:
 ---------------------